finish rearranging things
