+++ azure_data_cosmos_engine/src/error.rs
@@ -27,7 +27,7 @@ pub enum ErrorKind {
     /// Indicates that the query plan requires features that are not supported by the query engine.
     ///
     /// This error is not recoverable, and should be very rare (or even impossible).
-    /// The [`SUPPORTED_FEATURES_STRING`](crate::query::SUPPORTED_FEATURES_STRING) constant reports the features supported by the engine, and the language binding must provide that information to the gateway when generating a query plan.
+    /// The [`SUPPORTED_FEATURES`](crate::query::SUPPORTED_FEATURES) constant reports the features supported by the engine, and the language binding must provide that information to the gateway when generating a query plan.
     /// The gateway will return an error if the query requires features not listed in the supported features.
     /// We provide this error to cover cases where the language binding is incorrectly reporting the supported features, or edge cases where the engine is not correctly reporting the features it supports.
     UnsupportedQueryPlan,
